feat [server]: add basic user service validation and exception throwing

+++ server/src/modules/user/service/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+    ConflictException,
+    Injectable,
+    NotFoundException,
+} from '@nestjs/common';
 import { UpdateUserDto } from '../dto/update-user.dto';
 import { User } from '../entity/user.entity';
 import { UserRepository } from '../repository/user.repository';
@@ -8,6 +12,15 @@ export class UserService {
     constructor(private readonly userRepository: UserRepository) {}
 
     async create(user: User) {
+        const usernameExists = await this.userRepository.getByUsername(
+            user.username,
+        );
+        if (usernameExists)
+            throw new ConflictException('Username already exists');
+
+        const emailExists = await this.userRepository.getByEmail(user.email);
+        if (emailExists) throw new ConflictException('Email already exists');
+
         const result = await this.userRepository.create(user);
         return result;
     }
@@ -33,8 +46,11 @@ export class UserService {
     }
 
     // TODO: Check if it's ok to refresh token through here
+    // TODO: Implement username and email update restrictions
     async update(id: string, updateUserDto: UpdateUserDto) {
         const user = await this.userRepository.getById(id);
+        if (!user) throw new NotFoundException('User not found');
+
         const { ...updatedUserData } = updateUserDto;
         const updatedUser: User = {
             id: id,
@@ -51,6 +67,10 @@ export class UserService {
     }
 
     async delete(id: string) {
+        // check if user exists
+        const user = await this.userRepository.getById(id);
+        if (!user) throw new NotFoundException('User not found');
+
         const result = await this.userRepository.delete(id);
         return result;
     }
